# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.StockMoneyFlow:
    properties:
      created_at:
        type: string
      id:
        type: integer
      large_order_ratio:
        description: 大单成交占比(%)
        type: number
      main_net:
        description: 主力净流入
        type: number
      medium_order_ratio:
        description: 中单成交占比(%)
        type: number
      retail_net:
        description: 散户净流入
        type: number
      small_order_ratio:
        description: 小单成交占比(%)
        type: number
      symbol:
        description: 股票代码
        type: string
      trade_date:
        description: 交易日期
        type: string
    type: object
  models.StockMoneyFlowAll:
    properties:
      change_percent:
        type: number
      created_at:
        type: string
      id:
        type: integer
      inflow_amount:
        type: number
      latest_price:
        type: number
      name:
        type: string
      net_amount:
        type: number
      outflow_amount:
        type: number
      serial_number:
        type: integer
      symbol:
        type: string
      time_span:
        type: integer
      trade_date:
        type: string
      turnover:
        type: number
      turnover_rate:
        type: number
    type: object
host: 192.168.3.99:3003
info:
  contact: {}
  description: 股票基础信息管理 RESTful API
  title: 股票信息 API
  version: "1.0"
paths:
...
  /stock-money-flow:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StockMoneyFlow'
            type: array
      summary: 获取全部资金流数据
      tags:
      - 股票资金流
    post:
      consumes:
      - application/json
      parameters:
      - description: 股票资金流数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockMoneyFlow'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StockMoneyFlow'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 创建资金流数据
      tags:
      - 股票资金流
  /stock-money-flow-all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StockMoneyFlowAll'
            type: array
      summary: 获取全部资金流数据
      tags:
      - 股票资金流(ALL)
    post:
      consumes:
      - application/json
      parameters:
      - description: 股票资金流数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockMoneyFlowAll'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StockMoneyFlowAll'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 创建资金流数据
      tags:
      - 股票资金流(ALL)
  /stock-money-flow-all/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 删除资金流数据
      tags:
      - 股票资金流(ALL)
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockMoneyFlowAll'
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据ID获取资金流数据
      tags:
      - 股票资金流(ALL)
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 股票资金流数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockMoneyFlowAll'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockMoneyFlowAll'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 更新资金流数据
      tags:
      - 股票资金流(ALL)
  /stock-money-flow-all/symbol/{symbol}:
    get:
      parameters:
      - description: 股票代码
        in: path
        name: symbol
        required: true
        type: string
      - description: 交易日期(YYYY-MM-DD)
        in: query
        name: trade_date
        required: true
        type: string
      - description: 时间跨度(0,3,5,10)
        in: query
        name: time_span
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StockMoneyFlowAll'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据symbol和trade_date获取资金流数据
      tags:
      - 股票资金流(ALL)
  /stock-money-flow/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 删除资金流数据
      tags:
      - 股票资金流
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockMoneyFlow'
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据ID获取资金流数据
      tags:
      - 股票资金流
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 股票资金流数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockMoneyFlow'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockMoneyFlow'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 更新资金流数据
      tags:
      - 股票资金流
  /stock-money-flow/symbol/{symbol}:
    get:
      parameters:
      - description: 股票代码
        in: path
        name: symbol
        required: true
        type: string
      - description: 交易日期(YYYY-MM-DD)
        in: query
        name: trade_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockMoneyFlow'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据symbol和trade_date获取资金流数据
      tags:
      - 股票资金流
...
swagger: "2.0"